Move classes Node inherits from into a submodule
