+++ pyiron_workflow/io.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from typing import Any
-import warnings
 
 from pyiron_snippets.logger import logger
 from pyiron_snippets.dotdict import DotDict
@@ -25,8 +24,8 @@ from pyiron_workflow.channels import (
     AccumulatingInputSignal,
     NOT_DATA,
 )
-from pyiron_workflow.has_interface_mixins import HasChannel, HasLabel, HasRun, UsesState
-from pyiron_workflow.has_to_dict import HasToDict
+from pyiron_workflow.mixin.has_interface_mixins import HasChannel, HasLabel, HasRun, UsesState
+from pyiron_workflow.mixin.has_to_dict import HasToDict
 
 
 class IO(HasToDict, ABC):
