+++ pyiron_workflow/nodes/static_io.py
@@ -6,9 +6,9 @@ from typing import Optional, Literal, TYPE_CHECKING
 from pandas import DataFrame
 
 from pyiron_workflow.channels import InputData
-from pyiron_workflow.injection import OutputsWithInjection, OutputDataWithInjection
+from pyiron_workflow.mixin.injection import OutputsWithInjection, OutputDataWithInjection
 from pyiron_workflow.io import Inputs
-from pyiron_workflow.preview import HasIOPreview
+from pyiron_workflow.mixin.preview import HasIOPreview
 from pyiron_workflow.node import Node
 
 if TYPE_CHECKING:
